+++ src/App/VRChatContentPublisher.App/ViewModels/Pages/SettingsPageViewModel.cs
@@ -10,13 +10,15 @@ public sealed partial class SettingsPageViewModel(
     ConnectSettingsViewModel connectSettingsViewModel,
     SessionsSettingsViewModel sessionsSettingsViewModel,
     AboutSettingsViewModel aboutSettingsViewModel,
-    HttpProxySettingsViewModel httpProxySettingsViewModel) : PageViewModelBase
+    HttpProxySettingsViewModel httpProxySettingsViewModel,
+    DebugSettingsViewModel debugSettingsViewModel) : PageViewModelBase
 {
     public AccountsSettingsViewModel AccountsSettingsViewModel { get; } = accountsSettingsViewModel;
     public ConnectSettingsViewModel ConnectSettingsViewModel { get; } = connectSettingsViewModel;
     public SessionsSettingsViewModel SessionsSettingsViewModel { get; } = sessionsSettingsViewModel;
     public HttpProxySettingsViewModel HttpProxySettingsViewModel { get; } = httpProxySettingsViewModel;
     public AboutSettingsViewModel AboutSettingsViewModel { get; } = aboutSettingsViewModel;
+    public DebugSettingsViewModel DebugSettingsViewModel { get; } = debugSettingsViewModel;
     
     [RelayCommand]
     private void NavigateToHome()